file_utils: add a new function to generate test summary
GenerateTestSummary generates the test summary based on factory log. The
test summary contains (1) the failed test items and the failed reason
(2) the start and end time of all tests and (3) Goofy init count.
With (1) and (2), it is easier for readers to understand the test
running status and time. (3) often indicates reboot event.

BUG=b:265757602
TEST=make test & run on DUT

Change-Id: Ia0e459225f5caac53c62f67407730d2e0070e180
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4478852
Commit-Queue: Phoebe Wang <phoebewang@chromium.org>
Tested-by: Phoebe Wang <phoebewang@chromium.org>
Reviewed-by: Wei-Shuo Su <stevesu@google.com>