py/hwid: add the utility to check feature enablement permission
It adds a field in feature matching payload to instruct the factory
toolkit whether device of the given brand code is permitted for
versioned feature enablement.  The field is currently default to
`MIXED` so that the check can always pass before the actual data
is piped to the upstream service.

In addition, it introduces "default" spec in the payload so that
the server can control the matching logic (to use brand-specific
specs or shared spec) by preparing different payloads.

BUG=b:275014916
TEST=make test
TEST=./deploy/cros_hwid_service.sh test

Change-Id: I90eae868366ec257ea6b7b006838570ae2ed6614
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4483296
Tested-by: Yong Hong <yhong@chromium.org>
Reviewed-by: Wei-Shuo Su <stevesu@google.com>
Commit-Queue: Yong Hong <yhong@chromium.org>