arg_utils: Make arg_utils compatible with python3.8
After the recent python3.8 uprev in CL:4482641,
trying to check for non-Enum objects in Enum classes
will raise a TypeError.
So we modify the conditional expressions to make arg_utils compatible
with python3.8.
We also refactor the unittest of arg_utils for making each test case
focus on the function that they want to test, and add more test cases
for Arg() function.

BUG=b:279860710
TEST=(1)`make test` with python3.6 and python3.8
     (2)Manually run tests on corsola device and
        make sure it work as expected.

Change-Id: Iafc39da3c6e707980d37da3a9aa435aa36ebd7bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4506090
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Tested-by: LIU KUAN-FU <iamjeffliu@google.com>
Commit-Queue: LIU KUAN-FU <iamjeffliu@google.com>