pytests: delete gps.py
The test was designed for Android GPS testing. See
go/cros-factory-gps-android. The pytest has not been used for several
years. It depends on android software stack and also LitePoint fixture.
If someday in the future we need to do similar factory test, we probably
need quite amount of modification / redesign.  In this case, having the
design doc and a snapshot of the source code is enough for future
references. Keeping the dead code in the repository has very little
benefits.

BUG=None
TEST=make test

Change-Id: I7d668cc5057cf1e4c8ed0b95ea2c6696b8484683
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4505413
Auto-Submit: Stimim Chen <stimim@chromium.org>
Tested-by: Stimim Chen <stimim@chromium.org>
Commit-Queue: Cheng Yueh <cyueh@chromium.org>
Reviewed-by: Cheng Yueh <cyueh@chromium.org>