hwid/service: Update battery probe statements skip rule
Skip generating probe statements for battery components only if
technology of one of the batteries is Li-ion or Li-poly while comparing
two batteries. This is to avoid the skip for batteries like this:

```
battery1:
    manufacturer: mfr123456
    model_name: XL
    technology: LiP
battery2:
    manufacturer: mfr12345678
    model_name: XL
    technology: LiP
```

Because both the technologies are not common technology, the generated
probe statements would be exact match, and we should not skip them if
both of them are exact match.

BUG=b:277701772
TEST=./deploy/cros_hwid_service.sh test

Change-Id: I84364c66a7efc6cd5e861d93a0f07f9cd369ed47
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4505724
Tested-by: Allen Shih <allenshihmc@chromium.org>
Reviewed-by: Clark Chung <ckclark@chromium.org>
Commit-Queue: Allen Shih <allenshihmc@chromium.org>