cros_payload: skip copying partition if the size of DST is 1
The size of partition 6 of the GPT table is only 1. However, partition 6
of the release image is now 200MB [1]. This breaks cros_payload.sh since
it tries to install partition 6 from release image to the DUT.
Since size of 1 on GPT table usually means that the partition is only a
placeholder, we can skip copying it.

[1] https://crrev.com/c/4422049

BUG=b:280363878, b:280924270
TEST=run factory shim install

Change-Id: I5e0e7ec816c99a672127d1d55e96a961d3d3a59e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4508734
Tested-by: Phoebe Wang <phoebewang@chromium.org>
Auto-Submit: Phoebe Wang <phoebewang@chromium.org>
Reviewed-by: Chih-Yao Chuang <jasonchuang@google.com>
Commit-Queue: Chih-Yao Chuang <jasonchuang@google.com>