gooftool: Rename upload method "shopfloor" to "factory_serer"
This CL renames the upload method "shopfloor" to "factory_server" for
gooftool. This is because the method "shopfloor" is actually uploading
reports to umpire server (factory server), and the method name is making
partners confused.

This CL deprecates the upload method "shopfloor" rather than removes it.
Warnings are added to logs when users are still using the method
"shopfloor".

This CL updates the generic_common and generic_grt test lists to use the
renamed method "factory_server".

This CL changes the unittest function name in finalize_unittest to
reflect the renaming.

BUG=b:281573026
BUG=b:281773658
TEST=make test
TEST=Run `gooftool upload_report --help` on DUT and verify both
     `shopfloor` and `factory_server` methods are listed
TEST=Run `gooftool upload_report
     --upload_method=shopfloor:umpire_server_url#sn` on DUT and verify
     warning messages are printed and umpire server receives report
TEST=Run `gooftool upload_report
     --upload_method=factory_server:umpire_server_url#sn` on DUT and
     verify umpire server receives report
TEST=Run finalize via factory toolkit GUI on DUT and verify upload
     method is set to `factory_server` for pytest `Finalize`

Change-Id: I70407bf77664a0388625c097c1c7c7d1eb21ceee
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4520475
Commit-Queue: Wen-Ting Tseng <wdzeng@chromium.org>
Reviewed-by: Stimim Chen <stimim@chromium.org>
Tested-by: Wen-Ting Tseng <wdzeng@chromium.org>