py: Make input_http unittests compatible with Python3.8
In CL:4482641, Python3.8 was upgraded.
This change removed the redundant trailing comma in repr() of
BaseException and all exceptions that don't override __repr__).
It caused input_http_unittest.py and input_http_testlog_unittest.py
to fail.

To fix this, we changed to use assertRegex() for matching test result.
It makes the unittests compatible with both Python3.6 and Python3.8.

BUG=b:279860710, b:281800471, b:281800472
TEST=`make test` with Python3.6 and Python3.8.

Change-Id: Ic16853e9c498698d3aedd9a57bf0f8784d8d48e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4531603
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Reviewed-by: Chun-Tsen Kuo <chuntsen@chromium.org>
Tested-by: LIU KUAN-FU <iamjeffliu@google.com>
Commit-Queue: LIU KUAN-FU <iamjeffliu@google.com>