instalog: Make buffer_xxx_file_unittests compatible with Python3.8
In CL:4482641, Python3.8 was upgraded.
This change changed the behavior of virtual memory control of Python,
which caused the result of `psutil.memory_info().vms`
to be different from Python3.6,
and made `buffer_simple_file_unittest` fail.

To make the unittests compatible with both Python3.6 and Python3.8,
we removed the `testNonConsumableEventsMemoryUsage` test case
from `buffer_xxx_file_unittest` since it was unstable
due to the changed behavior of virtual memory control of Python.
After discussion, we decided to turn this test into a benchmark.

BUG=b:279860710, b:281800141
TEST=(1)`make test` with Python3.6 and Python3.8.
     (2)`bin/factory_env buffer_simple_file_benchmark_test.py`
        with Python3.6 and Python3.8.

Change-Id: I68ba78c66b67ce3521020507e388866763236799
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4535816
Commit-Queue: LIU KUAN-FU <iamjeffliu@google.com>
Reviewed-by: Sung-Chi Li <lschyi@chromium.org>
Tested-by: LIU KUAN-FU <iamjeffliu@google.com>