gooftool: Don't verify SN bits in RMA finalize stage
We don't expect to set SN bits in RMA,
so we should not verify SN bits in RMA either.
Otherwise, it will cause an error in the RMA finalize stage.

To fix this, we disabled the `VerifySnBits()` function in RMA.
We also renamed the `verify-sn-bits` command to `verify_sn_bits`
to keep the naming consistent.

BUG=b:281229026
TEST=(1)make test
     (2)`gooftool verify_sn_bits` on corsola DUT
        to verify the conditionals work as expected.

Change-Id: Ice377f92a27f95473460f6a182082e22e1a4a942
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4535085
Reviewed-by: Yu-An Wang <wyuang@google.com>
Commit-Queue: LIU KUAN-FU <iamjeffliu@google.com>
Tested-by: LIU KUAN-FU <iamjeffliu@google.com>