copy_minios: selectively copy the minios part
miniOS A/B are signed with different recovery key, with A signed with
the old one and B signed with the new one. This test copies the miniOS
partition with the same recovery key version as the firmware and
overrides the other miniOS partition. This logic aligns with the logic
in chromeos-install in https://crrev.com/c/4263489.

BUG=b:282589300
TEST=Verified that `gooftool verify_keys` failed if copy_minios did
     not run.

Change-Id: I652d6f7ed03b8c0dc579a74d194e76568a17bf91
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4537218
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Commit-Queue: Phoebe Wang <phoebewang@chromium.org>
Tested-by: Phoebe Wang <phoebewang@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4546879
Commit-Queue: Cheng Yueh <cyueh@chromium.org>
Auto-Submit: Phoebe Wang <phoebewang@chromium.org>