hwid/service: expire BOM data caches
To prevent out-of-memory error in memorystore, we should set expiry to
these caches and set the config of redis to "volatile-lru"[1].  This CL
also handles the case that any of cached chunks expires which causes
UnpicklingError.

[1] volatile-lru will evict caches with expiry set in an LRU approach.
We do use expiry of caches at CL splitting process where the caches
should not be evicted, so we choose LRU which will not evict such
caches.

BUG=b:267677465
TEST=monitor that App Engine does not report OOM errors.

Change-Id: Ib090c80e9daf7cafaf6a13b96d75508429be4f17
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4554541
Reviewed-by: Yong Hong <yhong@chromium.org>
Tested-by: Clark Chung <ckclark@chromium.org>
Commit-Queue: Yong Hong <yhong@chromium.org>
Auto-Submit: Clark Chung <ckclark@chromium.org>