ap_ro_verification: Set wpsr by ap_wpsr tool when wp is enabled
If wp is enabled, we should set a non-zero mask to verify the wp
status. And we should also check that the zero mask is not set
unexpectedly.
Since we can get the correct wpsr now, remove the default wpsr "0 0".
Fix the bug when getting wp status.

BUG=b:284381564
TEST=Enable wp, gooftool ti50_set_spi_data, gsctool -az, gsctool -aB
And the result is pass.

Cq-Depend: chromium:4567975
Change-Id: I8e57576537772ad84f596b40ac7a94788de8bd1e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4576655
Commit-Queue: Chih-Yao Chuang <jasonchuang@google.com>
Auto-Submit: Chih-Yao Chuang <jasonchuang@google.com>
Reviewed-by: Phoebe Wang <phoebewang@chromium.org>
Tested-by: Chih-Yao Chuang <jasonchuang@google.com>