wireless_antenna: Fix variable overlapping and interface not brought up
The inner function `SwitchAntennaWiFiChip.SwitchAntenna._SwitchAntenna`
directly modifies the variable `success` declared in outer function, and
in this case the modification will not work. This CL rewrites the test
flow to fix this problem.

In the function `SwitchAntenna` the wireless interface may not be
brought up. This is because the inner function `_SwitchAntenna` may
directly raise an exception when maximal try count exceeds or when the
pytest encounters unexpected error. This CL adds a finally block to fix
this problem.

In the function `SwitchAntennaWifiChip.SwitchAntenna`, when the
validation fails, there will be no more retry. But then an exception
will be raised with the error message `Failed to set antenna for 10
tries`. This message does not reflect the real test flow. This CL
replaces logging a warning message with raising an exception to fix this
problem.

In the beginning of the function `SwitchAntennaWifiChip.SwitchAntenna`,
we bring down the wireless interface. And when `_SwitchAntenna` is going
to retry we do the same action. Therefore this CL moves it from
`SwitchAntenna` to the beginning of `_SwitchAntenna` to remove redundant
code.

BUG=b:222638799
TEST=Run wireless antenna pytest on DUT

Change-Id: If868014d8f3c0692202f7cbad6e1dc65bbeed1e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4600175
Reviewed-by: Yu-An Wang <wyuang@google.com>
Auto-Submit: Wen-Ting Tseng <wdzeng@chromium.org>
Commit-Queue: Yu-An Wang <wyuang@google.com>
Tested-by: Wen-Ting Tseng <wdzeng@chromium.org>