camera: Allow testing internal USB cameras by VID/PID
On some devices, the HDMI input functionality is implemented as an
internal USB video device. This change allows testing that
functionality. The other option would be to add the HDMI input device as
a front or rear facing "camera" in cros_config. But that seems wrong
since it's not actually a camera, and adding it to the config causes it
to be labelled as a front/rear camera in the UI.

BUG=b:281643335
TEST=HDMI loopback test on Constitution

Change-Id: I279151bd19b488629385fe88e7e0c08be146e51b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4606733
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Tested-by: Pablo Ceballos <pceballos@chromium.org>
Reviewed-by: TH Lin <t.h_lin@quanta.corp-partner.google.com>
Tested-by: TH Lin <t.h_lin@quanta.corp-partner.google.com>
Commit-Queue: Pablo Ceballos <pceballos@chromium.org>
Reviewed-by: Yu-An Wang <wyuang@google.com>
Reviewed-by: Stefan Adolfsson <sadolfsson@chromium.org>