finalize_bundle: extract rootkey from keyset
The image in firmware updater doesn't signed with the correct rootkey
for custom labels. Instead, the rootkey is resigned with the file under
keyset/ when executing chromeos-firmwareupdate.

Thus, we should extract rootkey from keyset/ instead of from the image
to get the correct rootkeys.

BUG=b:286772453, b:222430396
TEST=run extract_firmware_info on custom label model
TEST=test_image chromiumos_base_image.bin FactoryScriptTest

Change-Id: Ib920967ed3661d01be3b8fa8050a3e3e384a86f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4604130
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Tested-by: Yu-An Wang <wyuang@google.com>
Commit-Queue: Yu-An Wang <wyuang@google.com>