gsctool: Fix feature management flags lock status for Ti50
APRO verification pytest has a step to provision board ID into
Ti50 before normal `Cr50WriteFlashInfo`. This fails feature
management flag provision as in the original implementation
we're checking whether the board ID already set or not and
this is a Cr50-only behavior. According to b/275356839#comment2,
Ti50 is using `FactoryMode::is_enabled()` to check if we can
write feature flags into GSC. This CL refactors the original
implementation so that it can correctly provision feature flags
for both Cr50/Ti50.

BUG=b:286566945
TEST=make test
On Cr50 device with board ID already set, set feature device data to
(True, 1) and run `probe eval-function feature_management_flags` and verify it to be (False, 0). For Ti50, also set feature device data to (True, 1) and toggle between factory modes. For factory mode enabled, we should get (True, 1) and when it's disabled we should get (False, 0).

For end-to-end integration testing, it has to be done in factory
with unprovisioned GSC chips.

Change-Id: Ib8f7ee84f9b77dbcc7d4e45930b903fcc7547f61
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4604137
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Commit-Queue: Wei-Shuo Su <stevesu@google.com>
Tested-by: Wei-Shuo Su <stevesu@google.com>