pygpt: Create a base class for Partition class
This is a preparation step to create a fake partition in future CLs.
A fake partition is not tied to an actual image file and it can be used
to inject arbitrary content to a GPT partition. For implementation, the
fake partition can inherit from `PartitionBase` instead of `Partition`.

Also remove NAMES_ENCODING and NAMES_LENGTH as they are not used
anymore.

BUG=b:282892503
TEST=make test

Change-Id: I35fca6ee57623a187d2cc9bfa2a9bc3e01bd56a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4604706
Reviewed-by: Phoebe Wang <phoebewang@chromium.org>
Commit-Queue: Cheng-Han Yang <chenghan@chromium.org>
Tested-by: Cheng-Han Yang <chenghan@chromium.org>