hwid/v3: Add null comp for new firmware comp
If a new firmware component is included in an old database, the new
firmware component will be appended to the index 0 and break the
existing HWID decoding.

Append a null component at index 0 when new firmware component is
added so that the old HWID string won't be mis-decoded.

BUG=b:287182521
TEST=make test

Change-Id: I6d1587cf3115884a22991368d202605defe260f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4612688
Commit-Queue: Yu-An Wang <wyuang@google.com>
Reviewed-by: Yong Hong <yhong@chromium.org>
Tested-by: Yu-An Wang <wyuang@google.com>