bundle_creator: remove unnecessary permission granting
Remove the unnecessary read permission granting while creating a factory
bundle, which might raise error if the creator is not a user email.

The read permission is still granted to the user if they download the
created factory bundle from the frontend UI server.

BUG=b:264512542
TEST=(1) ./deploy/bundle_creator.sh test-docker
TEST=(1) Manually create a failed firestore document which should be
         a successful request.
     (2) ./deploy/bundle_creator.sh run-docker dev
     (3) ./deploy/bundle_creator.sh trigger-retry dev 1
     (4) Manually check the factory bundle is created.

Change-Id: I7ed9135d1e0337b542c963837fae0f17b41559f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4614824
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Commit-Queue: Liang-Chieh Chen <jamesqaq@chromium.org>
Tested-by: Liang-Chieh Chen <jamesqaq@chromium.org>