hwid/service: Revise battery probe statement generation rule
Ectool extended the return value length of battery fields from 11 to 31
while some existing values in HWID still have length 11. This would
cause the probe to fail if the probe statements happen to be an exact
match. To solve this issue while preventing from making the rule more
complex, this CL revises the battery component probe statement
generation rule:

For a battery, if its `model_name` and `manufacturer` are prefixes of
the corresponding field values of another battery in the same HWID
database, and at least one of the field values is different between the
two batteries, apply an exact match to the probe statement. Otherwise,
apply a prefix match.

BUG=b:281479050
TEST=./deploy/cros_hwid_service.sh test

Change-Id: I007ce9e963ea74c2b24af40cddf36b519bc7232d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4615152
Tested-by: Allen Shih <allenshihmc@chromium.org>
Commit-Queue: Allen Shih <allenshihmc@chromium.org>
Reviewed-by: Kevin Lin <kevinptt@chromium.org>