hwid/service: handle CQ count of CL which is over limit
If a CL is added to CQ too many times, it might be caused by not
properly handled cases (e.g. turns into an empty change while being
submitted to ToT).  This CL checks the CL before adding a CQ vote by
counting the 'Commit-Queue+2' in the CL messages.  If the count exceeds
to certain amount, the service will add a hashtag and add developers to
triage this issue.

BUG=b:284978321
TEST=deploy/cros_hwid_service.sh test

Change-Id: Ib172fcca336a8fdfc067eb721a3ac075f1a15996
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4648994
Reviewed-by: Yong Hong <yhong@chromium.org>
Commit-Queue: Clark Chung <ckclark@chromium.org>
Tested-by: Clark Chung <ckclark@chromium.org>