bundle_creator: add GetBundleErrorMessage api
Add a new api named `GetBundleErrorMessage` to get the specific error
message of a request.

It matches `project` with a queried document to make sure that a user
couldn't get any error message by just changing `doc_id` in the request.
The permission of accessing a `project` would be checked on the frontend
side.

BUG=b:284286967
TEST=(1) ./deploy/bundle_creator.sh deploy-appengine-v2 dev
     (2) Use the command to send GetBundleErrorMessage request and
         manually check for 3 cases.
TEST=(1) ./deploy/bundle_creator.sh test-appengine-v2

Change-Id: I0a2a599ac8e87078db10cc01226c28350369f245
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4653751
Commit-Queue: Liang-Chieh Chen <jamesqaq@chromium.org>
Reviewed-by: Yu-An Wang <wyuang@google.com>
Tested-by: Liang-Chieh Chen <jamesqaq@chromium.org>