accelerometer: add animation mode
Add visualization of the device status and animation for the rotation
instruction. Since the animation might not be intuitive, it is added as a new mode for the user to choose.
The user can still opt for the original test display.
The original test only provide images for the user to align
to, while the new test display  help the user to see whether the
device rotation has been captured by the sensor.

BUG=b:288021678
TEST=manual test on brya and kukui
TEST=make test

Change-Id: I1ae3114873bfb4e5d8b280b665aa652476d20e9c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4680645
Commit-Queue: Elly Chao <ellychao@google.com>
Tested-by: Elly Chao <ellychao@google.com>
Reviewed-by: Yu-An Wang <wyuang@google.com>
Reviewed-by: Jeff Lin <jeffulin@google.com>