hwid/service: remove skip_avl_check_changed flag
The skip_avl_check state in HWID DB is not stored, so we cannot derive
the skip_avl_check_changed field.  In addition, CL:4666546 is not yet
deployed to both FHC and HWID Server, so it's safe to make an
incompatible change in the proto definition.

BUG=b:289968231
TEST=None

Change-Id: Ic7f95789ae5274f21eb9c721a8bd47c1134e5716
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4671767
Tested-by: Clark Chung <ckclark@chromium.org>
Reviewed-by: Yong Hong <yhong@chromium.org>
Commit-Queue: Clark Chung <ckclark@chromium.org>
Auto-Submit: Clark Chung <ckclark@chromium.org>