hwid/service: patch bundle_uuids in change units
Change units is patched with internal DB after CL:4337302 but
bundle_uuids is missing in CompChange.  So uuids will be erased once the
user manually modify a component with bundle_uuids.

This CL adds bundle_uuids in CompChange so the uuids will be patched
back when the comp is touched.

BUG=b:290876563
TEST=deploy/cros_hwid_service.sh test

Change-Id: Iabca8b5629fd6287f7c5607f19bc11b4c26a6c6a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4678705
Reviewed-by: Clark Chung <ckclark@chromium.org>
Commit-Queue: Yu-An Wang <wyuang@google.com>
Tested-by: Yu-An Wang <wyuang@google.com>