gsctool: add support for Ti50 initial factory mode
Checking Ti50 initial factory mode is crucial for factory processes
like flashing feature management flags, but prior to CL:4618889,
we do not have ways to probe this flag in OS level so we used CCD
factory mode as an approximation. The CL fixes this undone task with
support of gsctool features added in CL:4618889.

BUG=b:288826585, b:283717415
TEST=Manually test on DUT with steps
1. Enter CCD factory mode
2. Run `gootool gsc_write_flash_info --factory_process=RMA`
and it should execute without issues.

Change-Id: I40e861e238d1d99e98e498b9d2f8cabcc16bbf59
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4703714
Reviewed-by: Chih-Yao Chuang <jasonchuang@google.com>
Tested-by: Wei-Shuo Su <stevesu@google.com>
Commit-Queue: Wei-Shuo Su <stevesu@google.com>