ti50_ap_ro_verification: Skip modifying wpsr after Finalize
The command `gsctool -aE <wpsr>` will fail when we run
Ti50APROVerification after finalize.
So we skip modifying wpsr when the chip is not in initial factory mode
and wpsr is provisioned.
It prevent from the fails of Ti50APROVerification and Finalize
when re-flow or RMA.
We also fix the unreachable code in `GSCTool.ParseWpsr` function,
and add more unit-tests for gsctool functions.

BUG=b:291682665, b:292047953, b:291542022
TEST=(1)make test
     (2)Manually test on corsola device and make sure it is functional.

Change-Id: I56d9d2e8a3a041ddeecfd9a4f1112dfbe70831cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4710356
Tested-by: LIU KUAN-FU <iamjeffliu@google.com>
Reviewed-by: Chih-Yao Chuang <jasonchuang@google.com>
Commit-Queue: LIU KUAN-FU <iamjeffliu@google.com>