image_tool: fix missing symbolic links
In the `bundle` command, modify `AddResource` function to add the
symbolic link files to a list instead of copying they to the temporary
bundle directory.

It still uses `tar` command to compress the bundle directory, but it
doesn't use bzip2 format for compressing at first.  Then it use
`tarfile` module to open the tar file for adding the symbolic links,
and use bzip2 to compress it in the end.

BUG=b:288236682
TEST=(1) make test
TEST=(1) Prepare a `manifest.yaml` for the original issue's case.
     (2) bin/factory_env py/tools/finalize_bundle.py manifest.yaml
     (3) Enter the folder created by `finalize_bundle`, run
         `py/tools/image_tool.py` with the same arguments as doing
	 archive in the previous step.
     (4) Decompress the generated archive to make sure that
         `setup/cgpt` and `setup/futility` are symbolic links.
     (5) Run `./setup/image_tool rma create -o rma.bin` outside chroot
         and make sure that test/release images are wrapped in.

Change-Id: Ifd8617c5b31b3e3efe51754726f6c8cfc7da961e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4716540
Reviewed-by: Stimim Chen <stimim@chromium.org>
Commit-Queue: Liang-Chieh Chen <jamesqaq@chromium.org>
Tested-by: Liang-Chieh Chen <jamesqaq@chromium.org>