camera: Fix the QR code scanning area
The QR code scanning rectangle area is drawn in `drawRect`. But it was
out of range in rear camera (this.flipImage is false). See b/293815160
comment#1 attachment for the screenshots.

We actually don't need the if statement to draw the area. This CL
removes the if statement to make the QR code scanning area consistent
in both front and rear camera.

BUG=b:293815160
TEST=run FrontCameraAssembleQR in tablet and normal mode on brya
TEST=run RearCameraAssembleQR in tablet and normal mode on brya

Change-Id: I916f783b25456e92c467b9fe1b3e181081810132
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4732158
Commit-Queue: Tzu-Min Sun <jimmysun@chromium.org>
Tested-by: Tzu-Min Sun <jimmysun@chromium.org>
Reviewed-by: Phoebe Wang <phoebewang@chromium.org>