venv: Allow to replace venv created out of chroot
In https://crrev.com/c/4846448/comments/1b5da70a_a2a7053e, a developer
accidentally created a python3.11 venv out of chroot by calling
`make doc`.

They cannot replace it with `make doc` in chroot because the behavior of
virtualenv in/out of chroot are different.

The `/path/to/venv/bin/python` created by out of chroot virtualenv is a
symbolic link to /usr/bin/python3 so the return value of
`./devtools/mk/sphinx.venv/bin/python --version` is python3.8 inside
chroot even if the venv version is python3.11.

Use --copies to make a copy of python instead of make a link so
`/path/to/venv/bin/python --version` returns the correct version of
python.

The virtualenv in chroot doesn't support --copies so we also replace
virtualenv with `python -m venv`, where venv is a built-in module in
Python 3 while virtualenv is a third-party tool.

BUG=b:277698374
TEST=(0) The following commands are done out of chroot.
     (1) rm -rf ./devtools/mk/sphinx.venv
     (2) ./devtools/mk/sphinx.sh
     (3) cros_sdk --working-dir=. ./devtools/mk/sphinx.sh
     (4) ./devtools/mk/sphinx.sh
     (5) With this CL, all pass.
     (6) Without this CL, fails at (3).
     (7) rm -rf ./devtools/mk/sphinx.venv
     (8) Run ./devtools/mk/sphinx.sh without this CL
     (9) Run `cros_sdk --working-dir=. ./devtools/mk/sphinx.sh`
         with this CL
     (A) See the venv is replaced

Change-Id: I7c397b34a18563eecd23f6a95f1cb9d8d6614e3d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4872456
Reviewed-by: Chih-Yao Chuang <jasonchuang@google.com>
Commit-Queue: Cheng Yueh <cyueh@chromium.org>
Tested-by: Cheng Yueh <cyueh@chromium.org>