hwid/service: Enable HWID DB self-service for storage assembly comps
This CL adds the pcie_emmc_storage_assembly_converter to check the
alignment of probe info and probed value of HWID DB in respect of
PCIE-EMMC storage assembly components.  In addition, in the case of a
storage assembly exposed as an NVMe storage, an additional converter is
added into storage_converter to ensure the coverage.

BUG=b:295261984
TEST=./deploy/cros_hwid_service.sh test

Change-Id: Id7c8348189e846d5020f2887340c9218b8233745
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4812408
Reviewed-by: Yong Hong <yhong@chromium.org>
Commit-Queue: Clark Chung <ckclark@chromium.org>
Tested-by: Clark Chung <ckclark@chromium.org>