test_ui_unittest: Set random seed in the test case
If we run the unittest as a module, the seed will not be set. Since
there's only one test case used random number, move random.seed(0) to
the test case.
We might still need to figure out if there's an issue in the ui in
the future.

BUG=b:297443326
TEST=make test

Change-Id: Ibe250df8e1ea8cab5ba93e738d882754925ac4ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4812846
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Commit-Queue: Cheng Yueh <cyueh@chromium.org>
Auto-Submit: Chih-Yao Chuang <jasonchuang@google.com>
Tested-by: Chih-Yao Chuang <jasonchuang@google.com>