gooftool: Fail VerifyCrosConfig if vpd and boot cache are out of sync
When we modify the custom_label_tag in vpd, `cros_config / brand-code`
is not updated until the next boot.

If we run GSCFinalize in this state, the RLZ may be wrong.

Another approach is that we just use the brand code from the release
image cros_config but we don't have time to do that.

BUG=b:295440702
TEST=(1) Provision sku and firmware correctly.
     (2) Assume that without label or label "${custom}" are valid.
     (3) vpd -d custom_label_tag
     (4) reboot
     (5) gooftool verify_cros_config -v 4
     (6) Pass
     (7) vpd -s custom_label_tag="${custom}"
     (8) Without this CL, passes.
     (9) With this CL, fails with "VPD updated without reboot."

Change-Id: I965a358e450614a696fbfc87d4e613365293f521
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4835840
Reviewed-by: Wei-Shuo Su <stevesu@google.com>
Tested-by: Cheng Yueh <cyueh@chromium.org>
Commit-Queue: Cheng Yueh <cyueh@chromium.org>