cutoff: Clear in the beginning of process end
After the process is finished, we ask operator to remove ac and
display a qrcode for custom factory to customize their process.
Clear in the beginning to make sure the instruction of "remove ac"
won't be clear. And we still need to clear for the qrcode.

BUG=b:294154976
TEST=modify cutoff.json and do action_b

Change-Id: I653c888bc9e6167a8f8fe3ee19362dfe53e3ea19
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4826228
Tested-by: Chih-Yao Chuang <jasonchuang@google.com>
Reviewed-by: Wen-Ting Tseng <wdzeng@chromium.org>
Auto-Submit: Chih-Yao Chuang <jasonchuang@google.com>
Commit-Queue: Wen-Ting Tseng <wdzeng@chromium.org>