probe_info_service: Support converting probe values to probe parameters
To generate probe parameter suggestions, we need to identify the
difference between the source probe parameters and the actual probe
results. This CL makes the `IProbeStatementConverter` able to convert
probe values to probe parameters, so that we can compare probe values
and probe parameters in the same format.

BUG=b:293377003
TEST=./deploy/probe_info_service.sh test

Change-Id: I4ad06ca4a92a665eca02c7e62ebbe8b4d1815f18
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4862922
Tested-by: Allen Shih <allenshihmc@chromium.org>
Commit-Queue: Allen Shih <allenshihmc@chromium.org>
Reviewed-by: Yong Hong <yhong@chromium.org>