ti50_ap_ro_verification: Check wpsr for PVT/MP devices
As the discussion in b:272574767,
there may be some security concerns for the DUTs with "0 0" wpsr.
The PVT/MP devices with "0 0" could pass AP RO verification,
but the factory test should ensure that the WPSR values are not "0 0"
for preventing possible security problems for user.

BUG=b:272574767
TEST=(1)`make test`
     (2)Manually test on DUT to make sure it work as expected.

Change-Id: Ic7ac657bfb531c45a76c5c774bc2588014704dc4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4851834
Tested-by: LIU KUAN-FU <iamjeffliu@google.com>
Reviewed-by: Chih-Yao Chuang <jasonchuang@google.com>
Commit-Queue: LIU KUAN-FU <iamjeffliu@google.com>
Reviewed-by: Jett Rink <jettrink@chromium.org>