probe: Fix _GetIwDevices() not working
The original implementation of parsing `iw dev` output is not
working in the case when we're seeing `Unnamed/non-netdev interface`
virtual node preesnted in the interface list. This CL fixes this by
performing a two-step query so that we can precisely check if it's
the targeted `iw_type` we're looking for regarding a given interface,
rather than checking this with a hacky way to iterate 2 elements
in the list at a time.

BUG=b:297821189
TEST=Manually tested on Rex P2 with `Unnamed/non-netdev interface`
interface presented in `iw dev` output, with steps as
1. `probe search cellular`. This is the original command that failed.
2. `probe -v probe` for general probe test.

Change-Id: Ic7eaddbdd0f1824524dc61cb562ded5b058de03c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4860971
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Commit-Queue: Wei-Shuo Su <stevesu@google.com>
Tested-by: Wei-Shuo Su <stevesu@google.com>
Reviewed-by: Tzu-Min Sun <jimmysun@chromium.org>