probe_info_service: Add suggestion messages for battery components
The battery component is a special case for generating suggestions.
Since the probe values for battery components may be truncated due to
uses of legacy image or firmware, an additional prefix check is needed
to confirm whether users should be advised to use the latest image and
firmware. This CL introduces the check and suggestion messages
specifically for battery components.

BUG=b:293377003
TEST=./deploy/probe_info_service.sh test

Change-Id: I7e022f9bce3e027f5a5cc8d1e71bae689acd035e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4876400
Tested-by: Allen Shih <allenshihmc@chromium.org>
Commit-Queue: Allen Shih <allenshihmc@chromium.org>
Reviewed-by: Yong Hong <yhong@chromium.org>