probe_info_service: Support probe parameter normalization
To generate probe parameter suggestions, we have to compare the probe
parameters from the client with the probe parameters reverted from the
probe values. However, the case of string values of the probe parameters
from the client might be uncertain. While some of the parameters are
case-sensitive and some are case-insensitive, we need to normalize the
probe parameters from the client to handle the case properly before we
do the comparison. This CL implements a method that normalizes the probe
parameters by converting them to the probe statement's input and
reverting them back to the probe parameters.

BUG=b:293377003
TEST=./deploy/probe_info_service.sh test

Change-Id: I46fb24a3227d4de32d2c211d30880d8ac4296367
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4907532
Reviewed-by: Yong Hong <yhong@chromium.org>
Commit-Queue: Allen Shih <allenshihmc@chromium.org>
Tested-by: Allen Shih <allenshihmc@chromium.org>