hwid/v3: don't update region pattern for init DB
We'd like to allow adding region before update by probe result, but
essesntial components will be missing in this case.  Thus, this CL
modify the logic that skips update pattern if the database is a initial
DB.

This CL also refactors the logic of update pattern that reuses the
public method.

BUG=b:301043098
TEST=make test

Change-Id: I18fc29426a5d4c44fe846b88d30fdf7675cfa739
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4900399
Tested-by: Yu-An Wang <wyuang@google.com>
Reviewed-by: Clark Chung <ckclark@chromium.org>
Commit-Queue: Yu-An Wang <wyuang@google.com>