probe_info_service: Align the mmc_host category name with Runtime Probe
The category name of converters must be identical to the categories in
Runtime Probe to parse the probe results. However, the converter name
for PCIe-eMMC storage bridges should not only include the category name
and probe function name `mmc_host` since it can be an SD card reader or
a storage bridge with different arguments. To solve this problem, this
CL introduces a new parameter `converter_name` to customize names for
converters, while keeping the converter category names aligned with
Runtime Probe.

BUG=b:293377003
TEST=./deploy/probe_info_service.sh test

Change-Id: I09babf3b03ad419a7cfa0f01e7dfa3e0dc7dd7eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4907531
Commit-Queue: Allen Shih <allenshihmc@chromium.org>
Tested-by: Allen Shih <allenshihmc@chromium.org>
Reviewed-by: Clark Chung <ckclark@chromium.org>