hwid/service: skip firmware CL when DB not found
Firmware record might contains "design" as project name which is not a
valid DLM device.

In the HWID server side, it is not able to tell whether this project
is a design or DLM device. When a project name is not found, we can
only skip instead of raising an internal error.

BUG=b:303011029
TEST=deploy/cros_hwid_service.sh test

Change-Id: I69b67404801f5f9f64dcd447c9c92c8b9ea6a708
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4909367
Commit-Queue: Yu-An Wang <wyuang@google.com>
Tested-by: Yu-An Wang <wyuang@google.com>
Reviewed-by: Clark Chung <ckclark@chromium.org>